get chainconfig from shardmgt contract
