+++ smartcontract/service/native/governance/param.go
@@ -399,146 +399,6 @@ func (this *WithdrawParam) Deserialize(r io.Reader) error {
 	return nil
 }
 
-type Configuration struct {
-	N                    uint32
-	C                    uint32
-	K                    uint32
-	L                    uint32
-	BlockMsgDelay        uint32
-	HashMsgDelay         uint32
-	PeerHandshakeTimeout uint32
-	MaxBlockChangeView   uint32
-}
-
-func (this *Configuration) Serialize(w io.Writer) error {
-	if err := utils.WriteVarUint(w, uint64(this.N)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize n error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.C)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize c error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.K)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize k error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.L)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize l error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.BlockMsgDelay)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize block_msg_delay error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.HashMsgDelay)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize hash_msg_delay error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.PeerHandshakeTimeout)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize peer_handshake_timeout error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.MaxBlockChangeView)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize max_block_change_view error: %v", err)
-	}
-	return nil
-}
-
-func (this *Configuration) Deserialize(r io.Reader) error {
-	n, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize n error: %v", err)
-	}
-	c, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize c error: %v", err)
-	}
-	k, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize k error: %v", err)
-	}
-	l, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize l error: %v", err)
-	}
-	blockMsgDelay, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize blockMsgDelay error: %v", err)
-	}
-	hashMsgDelay, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize hashMsgDelay error: %v", err)
-	}
-	peerHandshakeTimeout, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize peerHandshakeTimeout error: %v", err)
-	}
-	maxBlockChangeView, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize maxBlockChangeView error: %v", err)
-	}
-	if n > math.MaxUint32 {
-		return fmt.Errorf("n larger than max of uint32")
-	}
-	if c > math.MaxUint32 {
-		return fmt.Errorf("c larger than max of uint32")
-	}
-	if k > math.MaxUint32 {
-		return fmt.Errorf("k larger than max of uint32")
-	}
-	if l > math.MaxUint32 {
-		return fmt.Errorf("l larger than max of uint32")
-	}
-	if blockMsgDelay > math.MaxUint32 {
-		return fmt.Errorf("blockMsgDelay larger than max of uint32")
-	}
-	if hashMsgDelay > math.MaxUint32 {
-		return fmt.Errorf("hashMsgDelay larger than max of uint32")
-	}
-	if peerHandshakeTimeout > math.MaxUint32 {
-		return fmt.Errorf("peerHandshakeTimeout larger than max of uint32")
-	}
-	if maxBlockChangeView > math.MaxUint32 {
-		return fmt.Errorf("maxBlockChangeView larger than max of uint32")
-	}
-	this.N = uint32(n)
-	this.C = uint32(c)
-	this.K = uint32(k)
-	this.L = uint32(l)
-	this.BlockMsgDelay = uint32(blockMsgDelay)
-	this.HashMsgDelay = uint32(hashMsgDelay)
-	this.PeerHandshakeTimeout = uint32(peerHandshakeTimeout)
-	this.MaxBlockChangeView = uint32(maxBlockChangeView)
-	return nil
-}
-
-type PreConfig struct {
-	Configuration *Configuration
-	SetView       uint32
-}
-
-func (this *PreConfig) Serialize(w io.Writer) error {
-	if err := this.Configuration.Serialize(w); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize configuration error: %v", err)
-	}
-	if err := utils.WriteVarUint(w, uint64(this.SetView)); err != nil {
-		return fmt.Errorf("utils.WriteVarUint, serialize setView error: %v", err)
-	}
-	return nil
-}
-
-func (this *PreConfig) Deserialize(r io.Reader) error {
-	config := new(Configuration)
-	err := config.Deserialize(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize configuration error: %v", err)
-	}
-	setView, err := utils.ReadVarUint(r)
-	if err != nil {
-		return fmt.Errorf("utils.ReadVarUint, deserialize setView error: %v", err)
-	}
-	if setView > math.MaxUint32 {
-		return fmt.Errorf("setView larger than max of uint32")
-	}
-	this.Configuration = config
-	this.SetView = uint32(setView)
-	return nil
-}
-
 type GlobalParam struct {
 	CandidateFee uint64 //unit: 10^-9 ong
 	MinInitStake uint32 //min init pos
