+++ smartcontract/service/native/governance/utils.go
@@ -25,92 +25,12 @@ import (
 
 	"github.com/ontio/ontology/smartcontract/service/native/ont"
 
-	"github.com/ontio/ontology-crypto/vrf"
 	"github.com/ontio/ontology/common"
-	"github.com/ontio/ontology/common/config"
-	vbftconfig "github.com/ontio/ontology/consensus/vbft/config"
 	cstates "github.com/ontio/ontology/core/states"
 	"github.com/ontio/ontology/smartcontract/service/native"
 	"github.com/ontio/ontology/smartcontract/service/native/auth"
 	"github.com/ontio/ontology/smartcontract/service/native/utils"
 )
-
-func GetPeerPoolMap(native *native.NativeService, contract common.Address, view uint32) (*PeerPoolMap, error) {
-	peerPoolMap := &PeerPoolMap{
-		PeerPoolMap: make(map[string]*PeerPoolItem),
-	}
-	viewBytes, err := utils.GetUint32Bytes(view)
-	if err != nil {
-		return nil, fmt.Errorf("getUint32Bytes, getUint32Bytes error: %v", err)
-	}
-	peerPoolMapBytes, err := native.CacheDB.Get(utils.ConcatKey(contract, []byte(PEER_POOL), viewBytes))
-	if err != nil {
-		return nil, fmt.Errorf("getPeerPoolMap, get all peerPoolMap error: %v", err)
-	}
-	if peerPoolMapBytes == nil {
-		return nil, fmt.Errorf("getPeerPoolMap, peerPoolMap is nil")
-	}
-	item := cstates.StorageItem{}
-	err = item.Deserialize(bytes.NewBuffer(peerPoolMapBytes))
-	if err != nil {
-		return nil, fmt.Errorf("deserialize PeerPoolMap error:%v", err)
-	}
-	peerPoolMapStore := item.Value
-	if err := peerPoolMap.Deserialize(bytes.NewBuffer(peerPoolMapStore)); err != nil {
-		return nil, fmt.Errorf("deserialize, deserialize peerPoolMap error: %v", err)
-	}
-	return peerPoolMap, nil
-}
-
-func putPeerPoolMap(native *native.NativeService, contract common.Address, view uint32, peerPoolMap *PeerPoolMap) error {
-	bf := new(bytes.Buffer)
-	if err := peerPoolMap.Serialize(bf); err != nil {
-		return fmt.Errorf("serialize, serialize peerPoolMap error: %v", err)
-	}
-	viewBytes, err := utils.GetUint32Bytes(view)
-	if err != nil {
-		return fmt.Errorf("getUint32Bytes, get viewBytes error: %v", err)
-	}
-	native.CacheDB.Put(utils.ConcatKey(contract, []byte(PEER_POOL), viewBytes), cstates.GenRawStorageItem(bf.Bytes()))
-	return nil
-}
-
-func GetGovernanceView(native *native.NativeService, contract common.Address) (*GovernanceView, error) {
-	governanceViewBytes, err := native.CacheDB.Get(utils.ConcatKey(contract, []byte(GOVERNANCE_VIEW)))
-	if err != nil {
-		return nil, fmt.Errorf("getGovernanceView, get governanceViewBytes error: %v", err)
-	}
-	governanceView := new(GovernanceView)
-	if governanceViewBytes == nil {
-		return nil, fmt.Errorf("getGovernanceView, get nil governanceViewBytes")
-	} else {
-		value, err := cstates.GetValueFromRawStorageItem(governanceViewBytes)
-		if err != nil {
-			return nil, fmt.Errorf("getGovernanceView, deserialize from raw storage item err:%v", err)
-		}
-		if err := governanceView.Deserialize(bytes.NewBuffer(value)); err != nil {
-			return nil, fmt.Errorf("deserialize, deserialize governanceView error: %v", err)
-		}
-	}
-	return governanceView, nil
-}
-
-func putGovernanceView(native *native.NativeService, contract common.Address, governanceView *GovernanceView) error {
-	bf := new(bytes.Buffer)
-	if err := governanceView.Serialize(bf); err != nil {
-		return fmt.Errorf("serialize, serialize governanceView error: %v", err)
-	}
-	native.CacheDB.Put(utils.ConcatKey(contract, []byte(GOVERNANCE_VIEW)), cstates.GenRawStorageItem(bf.Bytes()))
-	return nil
-}
-
-func GetView(native *native.NativeService, contract common.Address) (uint32, error) {
-	governanceView, err := GetGovernanceView(native, contract)
-	if err != nil {
-		return 0, fmt.Errorf("getView, getGovernanceView error: %v", err)
-	}
-	return governanceView.View, nil
-}
 
 func appCallTransferOnt(native *native.NativeService, from common.Address, to common.Address, amount uint64) error {
 	err := ont.AppCallTransfer(native, utils.OntContractAddress, from, to, amount)
@@ -227,137 +147,6 @@ func putGlobalParam2(native *native.NativeService, contract common.Address, glob
 		return fmt.Errorf("serialize, serialize globalParam2 error: %v", err)
 	}
 	native.CacheDB.Put(utils.ConcatKey(contract, []byte(GLOBAL_PARAM2)), cstates.GenRawStorageItem(bf.Bytes()))
-	return nil
-}
-
-func validatePeerPubKeyFormat(pubkey string) error {
-	pk, err := vbftconfig.Pubkey(pubkey)
-	if err != nil {
-		return fmt.Errorf("failed to parse pubkey")
-	}
-	if !vrf.ValidatePublicKey(pk) {
-		return fmt.Errorf("invalid for VRF")
-	}
-	return nil
-}
-
-func CheckVBFTConfig(configuration *config.VBFTConfig) error {
-	if configuration.C == 0 {
-		return fmt.Errorf("initConfig. C can not be 0 in config")
-	}
-	if int(configuration.K) != len(configuration.Peers) {
-		return fmt.Errorf("initConfig. K must equal to length of peer in config")
-	}
-	if configuration.L < 16*configuration.K || configuration.L%configuration.K != 0 {
-		return fmt.Errorf("initConfig. L can not be less than 16*K and K must be times of L in config")
-	}
-	if configuration.K < 2*configuration.C+1 {
-		return fmt.Errorf("initConfig. K can not be less than 2*C+1 in config")
-	}
-	if configuration.N < configuration.K || configuration.K < 7 {
-		return fmt.Errorf("initConfig. config not match N >= K >= 7")
-	}
-	if configuration.BlockMsgDelay < 5000 {
-		return fmt.Errorf("initConfig. BlockMsgDelay must >= 5000")
-	}
-	if configuration.HashMsgDelay < 5000 {
-		return fmt.Errorf("initConfig. HashMsgDelay must >= 5000")
-	}
-	if configuration.PeerHandshakeTimeout < 10 {
-		return fmt.Errorf("initConfig. PeerHandshakeTimeout must >= 10")
-	}
-	if configuration.MinInitStake < 10000 {
-		return fmt.Errorf("initConfig. MinInitStake must >= 10000")
-	}
-	if len(configuration.VrfProof) < 128 {
-		return fmt.Errorf("initConfig. VrfProof must >= 128")
-	}
-	if len(configuration.VrfValue) < 128 {
-		return fmt.Errorf("initConfig. VrfValue must >= 128")
-	}
-
-	indexMap := make(map[uint32]struct{})
-	peerPubkeyMap := make(map[string]struct{})
-	for _, peer := range configuration.Peers {
-		_, ok := indexMap[peer.Index]
-		if ok {
-			return fmt.Errorf("initConfig, peer index is duplicated")
-		}
-		indexMap[peer.Index] = struct{}{}
-
-		_, ok = peerPubkeyMap[peer.PeerPubkey]
-		if ok {
-			return fmt.Errorf("initConfig, peerPubkey is duplicated")
-		}
-		peerPubkeyMap[peer.PeerPubkey] = struct{}{}
-
-		if peer.Index <= 0 {
-			return fmt.Errorf("initConfig, peer index in config must > 0")
-		}
-		//check peerPubkey
-		if err := validatePeerPubKeyFormat(peer.PeerPubkey); err != nil {
-			return fmt.Errorf("invalid peer pubkey")
-		}
-		_, err := common.AddressFromBase58(peer.Address)
-		if err != nil {
-			return fmt.Errorf("common.AddressFromBase58, address format error: %v", err)
-		}
-	}
-	return nil
-}
-
-func getConfig(native *native.NativeService, contract common.Address) (*Configuration, error) {
-	config := new(Configuration)
-	configBytes, err := native.CacheDB.Get(utils.ConcatKey(contract, []byte(VBFT_CONFIG)))
-	if err != nil {
-		return nil, fmt.Errorf("native.CacheDB.Get, get configBytes error: %v", err)
-	}
-	if configBytes == nil {
-		return nil, fmt.Errorf("getConfig, configBytes is nil")
-	}
-	value, err := cstates.GetValueFromRawStorageItem(configBytes)
-	if err != nil {
-		return nil, fmt.Errorf("getConfig, deserialize from raw storage item err:%v", err)
-	}
-	if err := config.Deserialize(bytes.NewBuffer(value)); err != nil {
-		return nil, fmt.Errorf("deserialize, deserialize config error: %v", err)
-	}
-	return config, nil
-}
-
-func putConfig(native *native.NativeService, contract common.Address, config *Configuration) error {
-	bf := new(bytes.Buffer)
-	if err := config.Serialize(bf); err != nil {
-		return fmt.Errorf("serialize, serialize config error: %v", err)
-	}
-	native.CacheDB.Put(utils.ConcatKey(contract, []byte(VBFT_CONFIG)), cstates.GenRawStorageItem(bf.Bytes()))
-	return nil
-}
-
-func getPreConfig(native *native.NativeService, contract common.Address) (*PreConfig, error) {
-	preConfig := new(PreConfig)
-	preConfigBytes, err := native.CacheDB.Get(utils.ConcatKey(contract, []byte(PRE_CONFIG)))
-	if err != nil {
-		return nil, fmt.Errorf("native.CacheDB.Get, get preConfigBytes error: %v", err)
-	}
-	if preConfigBytes != nil {
-		preConfigStore, err := cstates.GetValueFromRawStorageItem(preConfigBytes)
-		if err != nil {
-			return nil, fmt.Errorf("getConfig, preConfigBytes is not available")
-		}
-		if err := preConfig.Deserialize(bytes.NewBuffer(preConfigStore)); err != nil {
-			return nil, fmt.Errorf("deserialize, deserialize preConfig error: %v", err)
-		}
-	}
-	return preConfig, nil
-}
-
-func putPreConfig(native *native.NativeService, contract common.Address, preConfig *PreConfig) error {
-	bf := new(bytes.Buffer)
-	if err := preConfig.Serialize(bf); err != nil {
-		return fmt.Errorf("serialize, serialize preConfig error: %v", err)
-	}
-	native.CacheDB.Put(utils.ConcatKey(contract, []byte(PRE_CONFIG)), cstates.GenRawStorageItem(bf.Bytes()))
 	return nil
 }
